+++ gluuengine/node/node.py
@@ -69,12 +69,7 @@ class DeployNode(object):
             cmd_list = [
                 "sudo wget {} -P /usr/bin".format(RECOVERY_SCRIPT),
                 "sudo chmod +x /usr/bin/recovery.py",
-                "sudo apt-get -qq install -y --force-yes supervisor python-pip",
-                "sudo pip -q install --upgrade pip",
-                "sudo pip -q install virtualenv",
-                "sudo mkdir -p /root/.virtualenvs",
-                "sudo virtualenv /root/.virtualenvs/recovery",
-                "sudo /root/.virtualenvs/recovery/bin/pip -q install jinja2",
+                "sudo apt-get -qq install -y --force-yes supervisor",
                 "sudo wget {} -P /etc/supervisor/conf.d".format(RECOVERY_CONF),
                 "sudo supervisorctl reload",
             ]